Migration to new default NUXT TS specs
