+++ components/item-list-nav.vue
@@ -9,7 +9,7 @@
 </template>
 
 <script lang="ts">
-import { Component, Prop, Vue } from "nuxt-property-decorator"
+import { Component, Prop, Vue } from "vue-property-decorator"
 
 @Component({})
 export default class ItemListNav extends Vue {
