+++ components/item.vue
@@ -25,7 +25,7 @@
 </template>
 
 <script lang="ts">
-import { Component, Prop, Vue } from "nuxt-property-decorator"
+import { Component, Prop, Vue } from "vue-property-decorator"
 
 import { timeAgo } from "~/plugins/filters"
