+++ components/spinner.vue
@@ -22,7 +22,7 @@
 </template>
 
 <script lang="ts">
-import { Component, Prop, Vue } from "nuxt-property-decorator"
+import { Component, Prop, Vue } from "vue-property-decorator"
 
 @Component({})
 export default class Spinner extends Vue {
