--- jsconfig.json
@@ -1,6 +1,0 @@
-{
-  "include": [
-    "./**/*",
-    "tsconfig.json"
-  ]
-}
